! I fixed a minor bug with the list_archive function. ! The CommentLoop class has been completely reworked and now allows themes to fetch the comment information itself and display it however they please with accessor methods (CommentLoop::id for the comment ID, CommentLoop::type for the comment type, and so on).
Note:
The comment system currently keeps track of the user who created the comment (if they're not a guest), however I have yet to make the comment system load the users current name yet. Also, the comment form should no longer show the Name, Email and Website inputs on the comment if they're logged in and should instead show "You are posting as ..." instead.